[Refactor] Relocate to `global.vo`
