+++ src/main/java/boombimapi/domain/place/dto/response/node/ViewportClusterNodeResponse.java
@@ -1,7 +1,7 @@
 package boombimapi.domain.place.dto.response.node;
 
 import boombimapi.domain.place.dto.type.MarkerType;
-import boombimapi.global.dto.Coordinate;
+import boombimapi.global.vo.Coordinate;
 import java.util.Map;
 
 public record ViewportClusterNodeResponse(
